fix: tests

+++ packages/atag/tests/teardown.js
@@ -5,16 +5,15 @@ const os = require('os');
 const path = require('path');
 
 const DIR = path.join(os.tmpdir(), 'jest_puppeteer_global_setup');
+const isJestWatchMode = !'JEST_WORKER_ID' in process.env;
 
-// Tear down webpack dev server in case it is alive.
-process.on('exit', () => {
-  if (global.__WEBPACK_DEV_SERVER__) {
+module.exports = async function() {
+  // In watch mode, do not kill webpack dev server each time.
+  if (!isJestWatchMode && global.__WEBPACK_DEV_SERVER__) {
     console.log(chalk.green('Teardown WebpackDevServer'));
     global.__WEBPACK_DEV_SERVER__.kill();
   }
-});
 
-module.exports = async function() {
   console.log(chalk.green('Teardown Puppeteer'));
   await global.__BROWSER_GLOBAL__.close();
   rimraf.sync(DIR);
